# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blog Platform Swagger",
    "version": "1.0.0",
    "description": "This is a sample server blog platform Swagger. You can access the Swagger JSON at [http://localhost:5000/swagger.json](http://localhost:5000/swagger.json)"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/searchblog": {
      "get": {
        "summary": "Search for a blog post by ID or title.",
        "description": "Retrieve a blog post by either its ID or title.",
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "description": "The ID of the blog post."
          },
          {
            "in": "query",
            "name": "title",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The title of the blog post."
          },
          {
            "name": "Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "Bearer <token>"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with the found blog post.",
            "content": {
              "application/json": {
                "example": {
                  "Status": "Success",
                  "message": "Search Blog Successful Add....!",
                  "data": {
                    "_id": "64cccd9935528903d686f66c",
                    "userName": "blog",
                    "title": "blog post",
                    "content": "This is the content of the blog post",
                    "status": "published",
                    "images": "image_91692340829860.jpg",
                    "password": 123
                  }
                }
              }
            }
          },
          "404": {
            "description": "Blog post not found.",
            "content": {
              "application/json": {
                "example": {
                  "Status": "FAILED",
                  "message": "Blog Not Found"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/searchcategories": {
      "get": {
        "summary": "Search for categories by ID or company name.",
        "description": "Retrieve categories by either their ID or company name.",
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "description": "The ID of the category."
          },
          {
            "in": "query",
            "name": "companyName",
            "schema": {
              "type": "string"
            },
            "description": "The name of the company associated with the category."
          },
          {
            "name": "Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "Bearer <token>"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with the found categories.",
            "content": {
              "application/json": {
                "example": {
                  "Status": "Success",
                  "message": "Search categories Successful Add....!",
                  "data": {
                    "_id": "64c0e7a3aed71535f108ed34",
                    "name": "Path",
                    "categories": "electronics",
                    "companyName": "vivo",
                    "productName": "tablets",
                    "modelNumber": "LKI5874WE44D5",
                    "size": "5.5 inch",
                    "price": 22000,
                    "password": 123
                  }
                }
              }
            }
          },
          "404": {
            "description": "Categories not found.",
            "content": {
              "application/json": {
                "example": {
                  "Status": "FAILED",
                  "message": "Categories Not Found"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
...
}